+++ pages/index.tsx
@@ -6,7 +6,7 @@ import prisma from "@/lib/prisma";
 import { GetServerSideProps } from "next";
 import NewTabLink from "@/components/NewTabLink";
 import { addItemDB, changePublishedDB, saveEditDB } from "@/utils/noteItemDB";
-import { NoteItem, changedNoteItems } from "@/utils/noteGroup";
+import { NoteItem, changedNoteItems } from "@/utils/noteItem";
 
 const initializedItem = (id: number): NoteItem => ({
   title: "-",
